+++ modules/tinymce/src/plugins/media/test/ts/browser/PlaceholderTest.ts
@@ -114,8 +114,7 @@ describe('browser.tinymce.plugins.media.core.PlaceholderTest', () => {
       editor.options.set('media_live_embeds', false);
     });
 
-    // TODO: TINY-4627/TINY-8382
-    it.skip('TBA: Set and assert script placeholder structure', () => pTestScriptPlaceholder(hook.editor(),
+    it('TBA: Set and assert script placeholder structure', () => pTestScriptPlaceholder(hook.editor(),
       '<p>\n' +
       '<script src="http://media1.tinymce.com/123456" type="text/javascript"></sc' + 'ript>\n' +
       '<script src="http://media2.tinymce.com/123456" type="text/javascript"></sc' + 'ript>\n' +
@@ -129,11 +128,10 @@ describe('browser.tinymce.plugins.media.core.PlaceholderTest', () => {
       placeholderStructure
     ));
 
-    // TODO: TINY-4627/TINY-8382
-    it.skip('TBA: Set and assert video placeholder structure', () => pTestPlaceholder(hook.editor(),
+    it('TBA: Set and assert video placeholder structure', () => pTestPlaceholder(hook.editor(),
       '/custom/video.mp4',
       '<p><video controls="controls" width="300" height="150">\n' +
-      '<source src="custom/video.mp4" type="video/mp4"></video></p>',
+      '<source type="video/mp4" src="custom/video.mp4"></video></p>',
       placeholderStructure
     ));